+++ publish/CHANGELOG.md
@@ -3,6 +3,7 @@
   - Adds a new field`objectSwap` to the `expand_rooms.yaml` file.
   - Adds support for custom room themes.
   - Changes the fields `objectSwap` and `objectData` in the `expand_locations.yaml` to not affect dungeons.
+  - Changes the field `noBuildDungeon` in the `expand_locations.yaml` to affect the whole zone when set to true.
   - Fixes custom dungeons not working.
   - Removes creator information automatically from blueprints and data values (so that things don't appear player built).